+++ .eslintrc.yaml
@@ -2,12 +2,16 @@ root: true
 env:
   es2021: true
   node: true
+rules:
+  "simple-import-sort/imports": error
+  "simple-import-sort/exports": error
 parser: "@typescript-eslint/parser"
 parserOptions:
   ecmaVersion: 2021
   sourceType: module
 plugins:
   - "@typescript-eslint"
+  - "simple-import-sort"
 extends:
   - "plugin:@typescript-eslint/recommended"
   - "htmlacademy/node"
