+++ package.json
@@ -12,7 +12,7 @@
     "start": "npm run build && node ./dist/main.rest.js",
     "start:dev": "nodemon",
     "build": "npm run clean && npm run compile",
-    "lint": "eslint src/ --ext .ts",
+    "lint": "eslint src/ --ext .ts --fix",
     "compile": "tsc -p tsconfig.json",
     "clean": "rimraf dist",
     "ts": "ts-node",
@@ -27,6 +27,7 @@
     "@typescript-eslint/parser": "6.7.0",
     "eslint": "8.49.0",
     "eslint-config-htmlacademy": "9.1.1",
+    "eslint-plugin-simple-import-sort": "10.0.0",
     "json-server": "0.17.3",
     "nodemon": "3.0.1",
     "pino-pretty": "10.2.0",
