+++ src/cli/generate.command.ts
@@ -1,10 +1,11 @@
-import {Command} from './command.interface.js';
-import {MockServerData} from '../shared/types/index.js';
-import {got} from 'got';
-import {TSVOfferGenerator} from '../shared/libs/offer-generator/index.js';
-import {getErrorMessage} from '../shared/helpers/index.js';
 import chalk from 'chalk';
+import {got} from 'got';
+
+import {getErrorMessage} from '../shared/helpers/index.js';
 import {TSVFileWriter} from '../shared/libs/file-writer/index.js';
+import {TSVOfferGenerator} from '../shared/libs/offer-generator/index.js';
+import {MockServerData} from '../shared/types/index.js';
+import {Command} from './command.interface.js';
 
 export class GenerateCommand implements Command {
   private readonly name = '--generate';
