+++ src/cli/help.command.ts
@@ -1,5 +1,6 @@
-import {Command} from './command.interface.js';
 import chalk from 'chalk';
+
+import {Command} from './command.interface.js';
 
 export class HelpCommand implements Command {
   private readonly name = '--help';
