+++ src/cli/version.command.ts
@@ -1,8 +1,10 @@
-import {Command} from './command.interface.js';
 import {readFileSync} from 'node:fs';
 import {resolve} from 'node:path';
-import {getErrorMessage} from '../shared/helpers/index.js';
+
 import chalk from 'chalk';
+
+import {getErrorMessage} from '../shared/helpers/index.js';
+import {Command} from './command.interface.js';
 
 type PackageJSONConfig = {
   version: string;
