+++ src/main.cli.ts
@@ -1,6 +1,7 @@
 #!usr/bin/env node
 import 'reflect-metadata';
-import {CLIApplication, VersionCommand, HelpCommand, ImportCommand, GenerateCommand} from './cli/index.js';
+
+import {CLIApplication, GenerateCommand,HelpCommand, ImportCommand, VersionCommand} from './cli/index.js';
 
 function bootstrap() {
   const cliApplication = new CLIApplication();
