+++ src/main.rest.ts
@@ -1,10 +1,12 @@
 import 'reflect-metadata';
+
 import {Container} from 'inversify';
+
 import {createRestApplicationContainer, RestApplication} from './rest/index.js';
-import {Components} from './shared/types/index.js';
-import {createUserContainer} from './shared/modules/user/index.js';
-import {createOfferContainer} from './shared/modules/offer/index.js';
 import {createCommentContainer} from './shared/modules/comment/index.js';
+import {createOfferContainer} from './shared/modules/offer/index.js';
+import {createUserContainer} from './shared/modules/user/index.js';
+import {Components} from './shared/types/index.js';
 
 async function bootstrap(): Promise<void> {
   const appContainer = Container.merge(
