+++ src/shared/libs/database-client/index.ts
@@ -1,3 +1,3 @@
+export * from './const.js';
 export * from './database-client.interface.js';
 export * from './mongo.database-client.js';
-export * from './const.js';
