+++ src/shared/libs/logger/console.logger.ts
@@ -1,5 +1,5 @@
-import {Logger} from './logger.interface.js';
 import {getErrorMessage} from '../../helpers/index.js';
+import {Logger} from './logger.interface.js';
 
 export class ConsoleLogger implements Logger {
   public debug(message: string, ...args:unknown[]): void{
