+++ src/shared/libs/logger/index.ts
@@ -1,3 +1,3 @@
+export * from './console.logger.js';
 export * from './logger.interface.js';
 export * from './pino.logger.js';
-export * from './console.logger.js';
