+++ src/shared/libs/logger/pino.logger.ts
@@ -1,8 +1,10 @@
-import {Logger as PinoInstance, pino, transport} from 'pino';
-import {Logger} from './logger.interface.js';
-import {getCurrentModuleDirectoryPath} from '../../helpers/index.js';
 import {resolve} from 'node:path';
+
 import {injectable} from 'inversify';
+import {Logger as PinoInstance, pino, transport} from 'pino';
+
+import {getCurrentModuleDirectoryPath} from '../../helpers/index.js';
+import {Logger} from './logger.interface.js';
 
 @injectable()
 export class PinoLogger implements Logger {
