+++ src/shared/libs/offer-generator/index.ts
@@ -1,3 +1,3 @@
+export * from './const.js';
 export * from './offer-generator.interface.js';
 export * from './tsv-offer-generator.js';
-export * from './const.js';
