+++ src/shared/libs/rest/controller/controller.interface.ts
@@ -1,5 +1,6 @@
 import {Response, Router} from 'express';
-import {Route} from '../types/route.interface.js';
+
+import {Route} from '../types/index.js';
 
 export interface Controller {
   readonly router: Router;
