+++ src/shared/libs/rest/controller/index.ts
@@ -1,3 +1,3 @@
-export * from './controller.interface.js';
 export * from './base-controller.abstaract.js';
 export * from './base-controller.constant.js';
+export * from './controller.interface.js';
