+++ src/shared/libs/rest/exception-filter/default-exception-filter.ts
@@ -1,12 +1,13 @@
-import {ExceptionFilter} from './exception-filter.interface.js';
-import {inject, injectable} from 'inversify';
-import {Components} from '../../../types/index.js';
-import {Logger} from '../../logger/index.js';
-import {Error} from 'mongoose';
 import {NextFunction, Request, Response} from 'express';
 import {StatusCodes} from 'http-status-codes';
-import {HttpError} from '../errors/index.js';
+import {inject, injectable} from 'inversify';
+import {Error} from 'mongoose';
+
 import {createErrorObject} from '../../../helpers/index.js';
+import {Components} from '../../../types/index.js';
+import {Logger} from '../../logger/index.js';
+import {HttpError} from '../errors/index.js';
+import {ExceptionFilter} from './exception-filter.interface.js';
 
 @injectable()
 export class DefaultExceptionFilter implements ExceptionFilter {
