+++ src/shared/libs/rest/exception-filter/index.ts
@@ -1,2 +1,2 @@
-export * from './exception-filter.interface.js';
 export * from './default-exception-filter.js';
+export * from './exception-filter.interface.js';
