+++ src/shared/libs/rest/index.ts
@@ -1,4 +1,4 @@
-export * from './types/index.js';
 export * from './controller/index.js';
-export * from './exception-filter/index.js';
 export * from './errors/index.js';
+export * from './exception-filter/index.js';
+export * from './types/index.js';
