+++ src/shared/libs/rest/types/index.ts
@@ -1,4 +1,4 @@
 export * from './http-methods.enum.js';
-export * from './route.interface.js';
 export * from './request-body.type.js';
 export * from './request-params.type.js';
+export * from './route.interface.js';
