+++ src/shared/modules/comment/comment-service.interface.ts
@@ -1,6 +1,7 @@
-import {CreateCommentDto} from './dto/create-comment.dto.js';
 import {DocumentType} from '@typegoose/typegoose';
+
 import {CommentEntity} from './comment.entity.js';
+import {CreateCommentDto} from './dto/create-comment.dto.js';
 
 export interface CommentService {
   create(dto: CreateCommentDto): Promise<DocumentType<CommentEntity>>;
