+++ src/shared/modules/comment/comment.container.ts
@@ -1,11 +1,12 @@
-import {Container} from 'inversify';
 import {types} from '@typegoose/typegoose';
-import {CommentEntity, CommentModel} from './comment.entity.js';
+import {Container} from 'inversify';
+
+import {Controller} from '../../libs/rest/index.js';
 import {Components} from '../../types/index.js';
+import {CommentController} from './comment.controller.js';
+import {CommentEntity, CommentModel} from './comment.entity.js';
 import {CommentService} from './comment-service.interface.js';
 import {DefaultCommentService} from './default-comment.service.js';
-import {Controller} from '../../libs/rest/index.js';
-import {CommentController} from './comment.controller.js';
 
 export function createCommentContainer() {
   const commentContainer = new Container();
