+++ src/shared/modules/comment/index.ts
@@ -1,6 +1,6 @@
+export * from './comment.container.js';
+export * from './comment.controller.js';
+export * from './comment.entity.js';
 export * from './comment-service.interface.js';
 export * from './default-comment.service.js';
-export * from './comment.container.js';
-export * from './comment.entity.js';
 export * from './dto/create-comment.dto.js';
-export * from './comment.controller.js';
