+++ src/shared/modules/offer/create-offer-request.type.ts
@@ -1,4 +1,5 @@
 import {Request} from 'express';
+
 import {RequestBody, RequestParams} from '../../libs/rest/index.js';
 import {CreateOfferDto} from './dto/create-offer.dto.js';
