+++ src/shared/modules/offer/default-offer.service.ts
@@ -1,12 +1,13 @@
-import {OfferService} from './offer-service.interface.js';
-import {CreateOfferDto} from './dto/create-offer.dto.js';
 import {DocumentType, types} from '@typegoose/typegoose';
-import {OfferEntity} from './offer.entity.js';
 import {inject, injectable} from 'inversify';
-import {Components, SortType} from '../../types/index.js';
+
 import {Logger} from '../../libs/logger/index.js';
+import {Components, SortType} from '../../types/index.js';
+import {CreateOfferDto} from './dto/create-offer.dto.js';
 import {UpdateOfferDto} from './dto/update-offer.dto.js';
 import {DEFAULT_OFFER_COUNT} from './offer.constant.js';
+import {OfferEntity} from './offer.entity.js';
+import {OfferService} from './offer-service.interface.js';
 
 @injectable()
 export class DefaultOfferService implements OfferService {
