+++ src/shared/modules/offer/index.ts
@@ -1,7 +1,7 @@
 export * from './default-offer.service.js';
-export * from './offer-service.interface.js';
-export * from './offer.container.js';
-export * from './offer.entity.js';
 export * from './dto/create-offer.dto.js';
 export * from './dto/update-offer.dto.js';
+export * from './offer.container.js';
 export * from './offer.controller.js';
+export * from './offer.entity.js';
+export * from './offer-service.interface.js';
