+++ src/shared/modules/offer/offer-service.interface.ts
@@ -1,7 +1,8 @@
-import {CreateOfferDto} from './dto/create-offer.dto.js';
 import {DocumentType} from '@typegoose/typegoose';
-import {OfferEntity} from './offer.entity.js';
+
+import {CreateOfferDto} from './dto/create-offer.dto.js';
 import {UpdateOfferDto} from './dto/update-offer.dto.js';
+import {OfferEntity} from './offer.entity.js';
 
 export interface OfferService {
   create(dto: CreateOfferDto): Promise<DocumentType<OfferEntity>>;
