+++ src/shared/modules/offer/offer.container.ts
@@ -1,11 +1,12 @@
+import {types} from '@typegoose/typegoose';
 import {Container} from 'inversify';
-import {OfferService} from './offer-service.interface.js';
+
+import {Controller} from '../../libs/rest/index.js';
 import {Components} from '../../types/index.js';
 import {DefaultOfferService} from './default-offer.service.js';
-import {types} from '@typegoose/typegoose';
-import {OfferEntity, OfferModel} from './offer.entity.js';
-import {Controller} from '../../libs/rest/index.js';
 import {OfferController} from './offer.controller.js';
+import {OfferEntity, OfferModel} from './offer.entity.js';
+import {OfferService} from './offer-service.interface.js';
 
 export function createOfferContainer() {
   const offerContainer = new Container();
