+++ src/shared/modules/user/create-user-request.type.ts
@@ -1,4 +1,5 @@
 import {Request} from 'express';
+
 import {RequestBody, RequestParams} from '../../libs/rest/index.js';
 import {CreateUserDto} from './dto/create-user.dto.js';
