+++ src/shared/modules/user/index.ts
@@ -1,7 +1,7 @@
-export * from './user.entity.js';
+export * from './default-user.service.js';
 export * from './dto/create-user.dto.js';
 export * from './dto/update-user.dto.js';
-export * from './default-user.service.js';
-export * from './user-service.interface.js';
 export * from './user.container.js';
 export * from './user.controller.js';
+export * from './user.entity.js';
+export * from './user-service.interface.js';
