+++ src/shared/modules/user/user-service.interface.ts
@@ -1,7 +1,8 @@
-import {CreateUserDto} from './dto/create-user.dto.js';
 import {DocumentType} from '@typegoose/typegoose';
-import {UserEntity} from './user.entity.js';
+
+import {CreateUserDto} from './dto/create-user.dto.js';
 import {UpdateUserDto} from './dto/update-user.dto.js';
+import {UserEntity} from './user.entity.js';
 
 export interface UserService {
   create(dto: CreateUserDto, salt: string): Promise<DocumentType<UserEntity>>;
