+++ src/shared/modules/user/user.container.ts
@@ -1,11 +1,12 @@
+import {types} from '@typegoose/typegoose';
 import {Container} from 'inversify';
-import {UserService} from './user-service.interface.js';
+
+import {Controller} from '../../libs/rest/index.js';
 import {Components} from '../../types/index.js';
 import {DefaultUserService} from './default-user.service.js';
-import {types} from '@typegoose/typegoose';
-import {UserEntity, UserModel} from './user.entity.js';
-import {Controller} from '../../libs/rest/index.js';
 import {UserController} from './user.controller.js';
+import {UserEntity, UserModel} from './user.entity.js';
+import {UserService} from './user-service.interface.js';
 
 export function createUserContainer() {
   const userContainer = new Container();
