+++ src/shared/types/index.ts
@@ -1,10 +1,10 @@
-export * from './offer.type.js';
-export * from './conveniences.enum.js';
 export * from './cities.enum.js';
+export * from './comment.interface.js';
+export * from './components.enum.js';
+export * from './conveniences.enum.js';
+export * from './mock-server-data.types.js';
+export * from './offer.type.js';
 export * from './places-types.enum.js';
+export * from './sort-type.enum.js';
 export * from './user.type.js';
 export * from './user-types.enum.js';
-export * from './mock-server-data.types.js';
-export * from './components.enum.js';
-export * from './comment.interface.js';
-export * from './sort-type.enum.js';
